+++ DietitianHub/src/main/java/com/dimstyl/dietitianhub/repositories/UserInfoRepository.java
@@ -4,6 +4,6 @@ import com.dimstyl.dietitianhub.entities.User;
 import com.dimstyl.dietitianhub.entities.UserInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserInfoRepository extends JpaRepository<UserInfo, Long> {
+public interface UserInfoRepository extends JpaRepository<UserInfo, Integer> {
     UserInfo findByUser(User user);
 }
